Fixed all review of previus commit
Fixed all review of previus commit.
Added two new files for HttpRequest, HttpResponse classes.
